+++ web/app/components/plugins/marketplace/atoms.ts
@@ -1,20 +1,30 @@
 import type { ActivePluginType } from './constants'
-import type { PluginsSort } from './types'
-import { atom, useAtom, useAtomValue } from 'jotai'
+import type { PluginsSort, SearchParamsFromCollection } from './types'
+import { atom, useAtom, useAtomValue, useSetAtom } from 'jotai'
 import { useQueryState } from 'nuqs'
+import { useCallback } from 'react'
 import { DEFAULT_SORT, PLUGIN_CATEGORY_WITH_COLLECTIONS } from './constants'
 import { marketplaceSearchParamsParsers } from './search-params'
 
-export const marketplaceSortAtom = atom<PluginsSort>(DEFAULT_SORT)
+const marketplaceSortAtom = atom<PluginsSort>(DEFAULT_SORT)
+export function useMarketplaceSort() {
+  return useAtom(marketplaceSortAtom)
+}
+export function useMarketplaceSortValue() {
+  return useAtomValue(marketplaceSortAtom)
+}
+export function useSetMarketplaceSort() {
+  return useSetAtom(marketplaceSortAtom)
+}
 
 /**
  * Preserve the state for marketplace
  */
 export const preserveSearchStateInQueryAtom = atom<boolean>(false)
 
-export const searchPluginTextAtom = atom<string>('')
-export const activePluginTypeAtom = atom<ActivePluginType>('all')
-export const filterPluginTagsAtom = atom<string[]>([])
+const searchPluginTextAtom = atom<string>('')
+const activePluginTypeAtom = atom<ActivePluginType>('all')
+const filterPluginTagsAtom = atom<string[]>([])
 
 export function useSearchPluginText() {
   const preserveSearchStateInQuery = useAtomValue(preserveSearchStateInQueryAtom)
@@ -51,4 +61,21 @@ export function useMarketplaceSearchMode() {
     || filterPluginTags.length > 0
     || (searchMode ?? (!PLUGIN_CATEGORY_WITH_COLLECTIONS.has(activePluginType)))
   return isSearchMode
+}
+
+export function useMarketplaceMoreClick() {
+  const [,setQ] = useSearchPluginText()
+  const setSort = useSetAtom(marketplaceSortAtom)
+  const setSearchMode = useSetAtom(searchModeAtom)
+
+  return useCallback((searchParams?: SearchParamsFromCollection) => {
+    if (!searchParams)
+      return
+    setQ(searchParams?.query || '')
+    setSort({
+      sortBy: searchParams?.sort_by || DEFAULT_SORT.sortBy,
+      sortOrder: searchParams?.sort_order || DEFAULT_SORT.sortOrder,
+    })
+    setSearchMode(true)
+  }, [setQ, setSort, setSearchMode])
 }
